chore: update bifold to 230
