+++ app/package.json
@@ -61,11 +61,11 @@
     "@formatjs/intl-relativetimeformat": "9.3.1",
     "@hyperledger/anoncreds-react-native": "0.2.1",
     "@hyperledger/aries-askar-react-native": "0.2.1",
-    "@hyperledger/aries-bifold-core": "1.0.0-alpha.222",
-    "@hyperledger/aries-bifold-remote-logs": "1.0.0-alpha.222",
-    "@hyperledger/aries-bifold-verifier": "1.0.0-alpha.222",
-    "@hyperledger/aries-oca": "1.0.0-alpha.222",
-    "@hyperledger/aries-react-native-attestation": "1.0.0-alpha.222",
+    "@hyperledger/aries-bifold-core": "1.0.0-alpha.230",
+    "@hyperledger/aries-bifold-remote-logs": "1.0.0-alpha.230",
+    "@hyperledger/aries-bifold-verifier": "1.0.0-alpha.230",
+    "@hyperledger/aries-oca": "1.0.0-alpha.230",
+    "@hyperledger/aries-react-native-attestation": "1.0.0-alpha.230",
     "@hyperledger/indy-vdr-react-native": "0.2.0",
     "@react-native-async-storage/async-storage": "1.15.11",
     "@react-native-community/masked-view": "0.1.11",
@@ -102,7 +102,7 @@
     "react-native-inappbrowser-reborn": "3.7.0",
     "react-native-keychain": "8.1.1",
     "react-native-localize": "2.2.6",
-    "react-native-logs": "^5.0.1",
+    "react-native-logs": "^5.1.0",
     "react-native-orientation-locker": "^1.6.0",
     "react-native-permissions": "4.0.1",
     "react-native-qrcode-svg": "6.2.0",
